Create an incomplete Mp4Box trait and derive macro
This version of the Mp4Box trait only has a size() item. In particular, it is
missing a getter for the box's type. Furthermore, size() does not include the
size of the type field. All of this should be fixed.

The proc-macro crate is of primary interest: the main library crate so far
contains only the Mp4Box trait declaration, and nonsense/fake box type
declarations used to test that the proc macro works. The latter should be
replaced with some (simple) actual box types from the standard once the proc
macro is functional enough to handle them.