WIP: mp4san-test: add remaining mandatory boxes
This is intended to be a fix for the failing tests.

It has progressed throught a few different failures, but it is still failing.
I'm also not sure if this is the intended solution. As far as I can tell, the
test failures began when more of the mp4 file format was added to mp4san's box
model (which must have been intended), and the test failures relate primarily
to mandatory boxes being missing—which does seem to suggest that it is.

Currently, the sanitizer runs to completion in the tests (which is progress, as
mentioned), but the tests subsequently fail because the sanitizer's output is
not as expected. Cursory examination of the input/output pairs suggests that
the sanitizer is rewriting box sizes—but not adding or removing any boxes—which
doesn't seem right to me, but I have not yet figured out what's going on there.